Reduce memory load

+++ kissim/comparison/feature_distances.py
@@ -156,8 +156,10 @@ class FeatureDistances:
         feature_distances = cls()
         feature_distances.structure_pair_ids = tuple(feature_distances_dict["structure_pair_ids"])
         feature_distances.kinase_pair_ids = tuple(feature_distances_dict["kinase_pair_ids"])
-        feature_distances.distances = np.array(feature_distances_dict["distances"])
-        feature_distances.bit_coverages = np.array(feature_distances_dict["bit_coverages"])
+        feature_distances.distances = np.array(feature_distances_dict["distances"], dtype=np.float)
+        feature_distances.bit_coverages = np.array(
+            feature_distances_dict["bit_coverages"], dtype=np.float
+        )
         return feature_distances
 
     def _get_feature_distances_and_bit_coverages(
